requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
Flask>=3.0.0
flask-cors>=4.0.0
dnspython>=2.4.2
python-whois>=0.9.4
scikit-learn>=1.3.0
numpy>=1.26.0
gunicorn>=21.2.0
